Day96/requirements.txt
Flask==2.3.2
stripe==6.13.1
requests==2.31.0
python-dotenv==1.0.